+++ src/instances/greathouse/app.ts
@@ -1,8 +1,0 @@
-import type { WorkbenchAppRegistration } from '../app-types'
-import GreathouseApp from './GreathouseApp.vue'
-import { greathouseInstance } from './config'
-
-export const workbenchAppRegistration = {
-  instanceId: greathouseInstance.id,
-  component: GreathouseApp,
-} satisfies WorkbenchAppRegistration