+++ src/instances/greathouse/instance.ts
@@ -1,8 +1,0 @@
-import type { WorkbenchInstanceRegistration } from '../instance-types'
-import { greathouseInstance } from './config'
-import { greathousePilotSnapshot } from './pilot'
-
-export const workbenchInstanceRegistration = {
-  instance: greathouseInstance,
-  staticSnapshot: greathousePilotSnapshot,
-} satisfies WorkbenchInstanceRegistration